Tweaked some tasks and moved TaskScrollCommands
