+++ src/main/java/rbasamoyai/industrialwarfare/client/screen/taskscroll/TaskScrollScreen.java
@@ -24,8 +24,8 @@ import rbasamoyai.industrialwarfare.IndustrialWarfare;
 import rbasamoyai.industrialwarfare.client.screen.selectors.ArgSelector;
 import rbasamoyai.industrialwarfare.client.screen.selectors.TaskScrollArgSelectorWidget;
 import rbasamoyai.industrialwarfare.common.containers.TaskScrollContainer;
+import rbasamoyai.industrialwarfare.common.entityai.taskscrollcmds.TaskScrollCommand;
 import rbasamoyai.industrialwarfare.common.items.taskscroll.TaskScrollOrder;
-import rbasamoyai.industrialwarfare.common.taskscrollcmds.TaskScrollCommand;
 import rbasamoyai.industrialwarfare.core.network.IWNetwork;
 import rbasamoyai.industrialwarfare.core.network.messages.STaskScrollSyncMessage;
 import rbasamoyai.industrialwarfare.utils.TextureUtils;
